requirements.txt
# app
fastapi==0.73.0
uvicorn==0.17.4
pydantic==1.9

python-dotenv==0.20.0

# db
SQLAlchemy==1.4.27
pymysql==1.0